+++ trunk/Rax/src/rax/dao/ibatis/ArticleContentDaoImpl.java
@@ -13,13 +13,13 @@ public class ArticleContentDaoImpl extends SqlMapClientDaoSupport implements
         ArticleContentDao {
 
     @Override
-    public Long create(String newInstance) throws DataAccessException {
-        return (Long) getSqlMapClientTemplate().insert("createArticleContent",
+    public int create(String newInstance) throws DataAccessException {
+        return (Integer) getSqlMapClientTemplate().insert("createArticleContent",
                 newInstance);
     }
 
     @Override
-    public String read(Long id) throws DataAccessException {
+    public String read(int id) throws DataAccessException {
         // TODO
         // Import multiple line into one
         return (String) getSqlMapClientTemplate().queryForObject(
@@ -41,19 +41,19 @@ public class ArticleContentDaoImpl extends SqlMapClientDaoSupport implements
     }
 
     @Override
-    public int deleteByCategoryId(Long id) throws DataAccessException {
+    public int deleteByCategoryId(int id) throws DataAccessException {
         return getSqlMapClientTemplate().delete(
                 "deleteArticleContentByCategoryId", id);
     }
 
     @Override
-    public int countByCategoryId(Long id) throws DataAccessException {
+    public int countByCategoryId(int id) throws DataAccessException {
         return (Integer) getSqlMapClientTemplate().queryForObject(
                 "countArticleContentByCategoryId", id);
     }
 
     @Override
-    public List<String> listByCategoryId(Long id, int index, int num)
+    public List<String> listByCategoryId(int id, int index, int num)
             throws DataAccessException {
         Map param = new HashMap();
         param.put("index", index);
@@ -64,7 +64,7 @@ public class ArticleContentDaoImpl extends SqlMapClientDaoSupport implements
     }
 
     @Override
-    public List<String> listAllByCategoryId(Long id) throws DataAccessException {
+    public List<String> listAllByCategoryId(int id) throws DataAccessException {
         return getSqlMapClientTemplate().queryForList(
                 "listAllArticleContentByCategoryId", id);
     }
